+++ src/main/java/com/github/patbattb/telegrambot/domain/TelegramBot.java
@@ -1,17 +1,23 @@
-package com.github.patbattb.telegrambot.service;
+package com.github.patbattb.telegrambot.domain;
 
+import com.github.patbattb.telegrambot.command.CommandMap;
+import com.github.patbattb.telegrambot.service.SendMessageBotService;
 import org.springframework.beans.factory.annotation.Value;
 import org.springframework.stereotype.Component;
 import org.telegram.telegrambots.bots.TelegramLongPollingBot;
 import org.telegram.telegrambots.meta.TelegramBotsApi;
-import org.telegram.telegrambots.meta.api.methods.send.SendMessage;
 import org.telegram.telegrambots.meta.api.objects.Update;
 import org.telegram.telegrambots.meta.exceptions.TelegramApiException;
 
+/**
+ * Telegram Bot with username and token from application.properties
+ */
 @Component
 public class TelegramBot extends TelegramLongPollingBot {
 
     private final String username;
+
+    private final CommandMap commandMap;
 
     public TelegramBot(
             @Value("${bot.username}") String username,
@@ -21,6 +27,7 @@ public class TelegramBot extends TelegramLongPollingBot {
     {
         super(token);
         this.username = username;
+        this.commandMap = new CommandMap(new SendMessageBotService(this));
         telegramBotsApi.registerBot(this);
     }
 
@@ -28,9 +35,8 @@ public class TelegramBot extends TelegramLongPollingBot {
     public void onUpdateReceived(Update update) {
         if (update.hasMessage() && update.getMessage().hasText()) {
             String message = update.getMessage().getText().trim();
-            long chatId = update.getMessage().getChatId();
 
-            sendMessage(chatId, message);
+            commandMap.getCommand(message).execute(update);
 
         }
     }
@@ -40,15 +46,4 @@ public class TelegramBot extends TelegramLongPollingBot {
         return username;
     }
 
-    private void sendMessage(long chatId, String message) {
-        var sm = new SendMessage();
-        sm.setChatId(chatId);
-        sm.setText(message);
-
-        try {
-            execute(sm);
-        } catch (TelegramApiException e) {
-            throw new RuntimeException(e);
-        }
-    }
 }