+++ src/main/java/com/github/patbattb/telegrambot/service/SendMessageBotService.java
@@ -1,0 +1,36 @@
+package com.github.patbattb.telegrambot.service;
+
+import com.github.patbattb.telegrambot.domain.TelegramBot;
+import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.stereotype.Service;
+import org.telegram.telegrambots.meta.api.methods.send.SendMessage;
+import org.telegram.telegrambots.meta.exceptions.TelegramApiException;
+
+/**
+ * Implementation of {@link SendMessageService} interface.
+ * Service for sending messages.
+ */
+@Service
+public class SendMessageBotService implements SendMessageService {
+
+    private final TelegramBot bot;
+
+    @Autowired
+    public SendMessageBotService(TelegramBot bot) {
+        this.bot = bot;
+    }
+
+    @Override
+    public void sendMessage(Long chatId, String message) {
+        var sendMessage = new SendMessage();
+        sendMessage.setChatId(chatId);
+        sendMessage.enableHtml(true);
+        sendMessage.setText(message);
+
+        try {
+            bot.execute(sendMessage);
+        } catch (TelegramApiException e) {
+            e.printStackTrace();
+        }
+    }
+}